Moved all errors and structs to be accessible from aiokafka namespace, without importing from kafka namespace.
