+++ aiokafka/errors.py
@@ -1,3 +1,23 @@
+from kafka.errors import *  # noqa
+from kafka.errors import KafkaError
+
+__all__ = [
+    # kafka-python errors
+    "KafkaError", "ConnectionError", "NodeNotReadyError",
+    "KafkaTimeoutError", "UnknownTopicOrPartitionError",
+    "UnrecognizedBrokerVersion", "NotLeaderForPartitionError",
+    "LeaderNotAvailableError", "TopicAuthorizationFailedError",
+    "OffsetOutOfRangeError", "MessageSizeTooLargeError", "for_code", "NoError",
+    "StaleMetadata", "CorrelationIdError", "NoBrokersAvailable",
+    "RebalanceInProgressError", "IllegalGenerationError",
+    "UnknownMemberIdError", "GroupLoadInProgressError",
+    "GroupCoordinatorNotAvailableError", "NotCoordinatorForGroupError",
+    "GroupAuthorizationFailedError",
+    # aiokafka custom errors
+    "ConsumerStoppedError", "NoOffsetForPartitionError", "RecordTooLargeError",
+    "ProducerClosed"
+]
+
 
 class ConsumerStoppedError(Exception):
     """ Raised on `get*` methods of Consumer if it's cancelled, even pending
@@ -10,3 +30,15 @@ class IllegalOperation(Exception):
         current configuration. For example trying to commit if no group_id was
         given.
     """
+
+
+class NoOffsetForPartitionError(KafkaError):
+    pass
+
+
+class RecordTooLargeError(KafkaError):
+    pass
+
+
+class ProducerClosed(KafkaError):
+    pass
